requirements.txt
aiosqlite
discord.py
profanity-check
requests
yt-dlp